requirements.txt
# requirements.txt
flwr
pytorch-lightning
pandas
numpy
streamlit
shap
matplotlib
scikit-learn 